exception handler 추가

+++ src/main/java/com/example/demo/repositories/MembersRepository.java
@@ -3,7 +3,5 @@ package com.example.demo.repositories;
 import com.example.demo.domain.Members;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface MembersRepository extends JpaRepository<Members, BigInteger> {
+public interface MembersRepository extends JpaRepository<Members, Long> {
 }
